% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{function for calculating one missing measurement of a food dispenser}
\usage{
calculate(rho = 0.6, kg = 25, angle = 45, depth = 40, width = NA,
  height = NA)
}
\arguments{
\item{rho}{Bulk density. By default 0.6 kg per liter because it will provide enough place for almost every cereal.}

\item{kg}{How much kilos shall the dispenser contain? By default 25 kg}

\item{angle}{Angle of the big slanted board inside the dispenser; unit is degree; by default 45}

\item{depth}{Dispenser's depth. Because of 'euro-norm' 40cm by default.}

\item{width}{Dispenser's widht. Either width or height is required! Measurements in cm.}

\item{height}{Dispenser's height.}
}
\value{
If width is given, height will be calculated. If height is given, width will be calculated. You can play with the angle input as well as the depth input. Output is a vector containing width, height, depth and angle.
}
\description{
function for calculating one missing measurement of a food dispenser
}
